refactor(iroh-blobs)!: make TempTag non-Clone (#2338)
## Description

refactor(iroh-blobs): Make TempTag non-Clone

I found that in most places it is not needed, and if somebody needs them
to be clone they can always wrap them in an Arc.

Also, this complicates extending the concept of temp tags via the rpc
boundary.

With this change if will be possible to use the same TempTag type both
in the low level blobs store API and in the higher level blobs API of
iroh.

## Breaking Changes

Changes the signature of TempTag::new to take a `Weak<dyn TagDrop>`
instead
of an Arc<dyn LivenessTracker>. This will affect projects that write
their own store
impl, so none that I am aware of. Nevertheless it is a breaking change.

The high level rpc API is not affected, since it does not know temp tags
(yet).

## Notes & open questions

Note: this is just part 1/x of changes needed to extend temp tags to the
rpc api. On its own it does not provide much value.

Note2: you could be even more radical and give a TempTag a lifetime and
a reference to its store. But I am pretty sure that this would not play
well with creating bindings and might even be inconvenient from pure
rust.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.