fix(iroh-net): prevent adding addressing info that points back to us (#2333)
## Description

Failed disco boxes for which the sender is our own node is are because
we tried to contact another node for which the addressing information we
have points back to us. We are the sender (this explains the source of
the disco box being ourselves) but we are not the meant recipient (thus
the failed decryption - we should always be able to decrypt a message
meant to ourselves - ).

This is of course undesirable. This Pr should mostly mitigate this
issue, but a failed disco box to ourselves can happen in other less
likely ways.

## Breaking Changes

n/a

## Notes & open questions

n/a

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] ~Tests if relevant.~
- [ ] ~All breaking changes documented.~