fix(iroh-gossip): connection loop misuses `tokio::select!` leading to read errors (#2572)
## Description

The connection loop of iroh-gossip misused tokio-select by selecting
over a future that is not cancellation safe. This means that if the
timings are bad, the message reading future would be aborted midway in a
message, and then restart by reading a length, which would then yield
some random number because it would be reading some random bytes in the
middle of a message. This means it would lead to random connection
drops.

## Breaking Changes

Backport from #2570 

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [ ] ~~Tests if relevant.~~
- [x] All breaking changes documented.